refactor(rtc): move RTT ingress and policy ownership
