+++ packages/shared-server/rallar-system/rtc-topology/topic/rtc-rtt-refinement-service.ts
@@ -28,6 +28,7 @@ type ExpiringDecision = Readonly<{
   expireAtEpochMs: number;
 }>;
 
+/** Coordinates stable refinement decisions across durable work retries. */
 export class RtcRttRefinementService {
   private readonly dependencies: RtcRttRefinementServiceDependencies;
   private readonly observations = new Map<string, ExpiringObservation>();
