+++ packages/shared-server/rallar-system/topology/planning/group-topology-planning-service.ts
@@ -18,7 +18,7 @@ import {
   type RallarRtcTopologyUpdateResult,
   type RtcTopologyPlanningIntent,
 } from '../../services/rallar-rtc-topology-service.ts';
-import { filterRtcRttMeasurementsForGroup } from '../../services/rtc-rtt-measurement-policy.ts';
+import { filterRtcRttMeasurementsForGroup } from '../../rtc-topology/policy/rtc-rtt-measurement-policy.ts';
 import { GroupTopologyValidationError } from '../group-topology-errors.ts';
 import type {
   GroupTopologyGroupSnapshotReader,
