+++ packages/tests/shared-server/ws-system-topics-rtc-topology.test.ts
@@ -17,16 +17,11 @@ import type { AuditStamp, GroupSnapshot } from '@shared/api/group-types.ts';
 import type { ClientSnapshot } from '@shared/api/client-types.ts';
 import * as clientStateSnapshotsRepository from '@shared/repository/client-state-snapshots-repository.ts';
 import * as groupStateSnapshotsRepository from '@shared/repository/group-state-snapshots-repository.ts';
-import { latestRttById } from '@shared/repository/rtt-repository.ts';
-import {
-  initRallarSystemWsTopics,
-  type InitRallarSystemWsTopicsOptions,
-} from '@shared-server/rallar-system/ws-system-topics.ts';
+import { initRallarSystemWsTopics } from '@shared-server/rallar-system/ws-system-topics.ts';
 import { createRtcTopologyOutboxPublisher } from '@shared-server/rallar-system/services/RtcTopologyOutboxWork.ts';
 import { RallarRtcTopologyService } from '@shared-server/rallar-system/services/rallar-rtc-topology-service.ts';
 import { GroupTopologyManagementService } from '@shared-server/rallar-system/topology/group-topology-management-service.ts';
 import { RtcTopologyExecutionRepository } from '@shared-server/rallar-system/repositories/RtcTopologyExecutionRepository.ts';
-import * as vivaldiService from '@shared-graph/vivaldi-service.ts';
 import { configureTestCacheRepositories } from '../cache-repository-config.ts';
 import { FakeRuntimeStateRepository } from './fake-runtime-state-repository.ts';
 import type { PSqlSql } from '@shared-server/postgres/PostgresSqlClient.ts';
@@ -208,341 +203,6 @@ describe('Rallar system websocket topics RTC topology', () => {
     expect(countSentTopologyMessages(sockets)).toBe(0);
   });
 
-  it('rejects RTT measurements from a mismatched AL sender', async () => {
-    configureTestCacheRepositories();
-    const { sockets } = createRttHarness(['session-a', 'session-b']);
-    const group = createGroupSnapshot('room-1', ['session-a', 'session-b']);
-    groupStateSnapshotsRepository.setGroupStateSnapshot(group);
-
-    await dispatchRtt(
-      sockets.get('session-a')!,
-      'session-a',
-      {
-        sessionIdFrom: 'session-b',
-        sessionIdTo: 'session-a',
-        rttMs: 12,
-        createdAtEpochMs: 1,
-        version: 1,
-      },
-      group,
-    );
-
-    expect(latestRttById().read('session-a::session-b')).toBeUndefined();
-  });
-
-  it('rejects RTT measurements for self pairs', async () => {
-    configureTestCacheRepositories();
-    const { sockets } = createRttHarness(['session-a']);
-    const group = createGroupSnapshot('room-1', ['session-a']);
-    groupStateSnapshotsRepository.setGroupStateSnapshot(group);
-
-    await dispatchRtt(
-      sockets.get('session-a')!,
-      'session-a',
-      {
-        sessionIdFrom: 'session-a',
-        sessionIdTo: 'session-a',
-        rttMs: 12,
-        createdAtEpochMs: 1,
-        version: 1,
-      },
-      group,
-    );
-
-    expect(latestRttById().read('session-a::session-a')).toBeUndefined();
-  });
-
-  it('rejects invalid RTT measurements', async () => {
-    configureTestCacheRepositories();
-    const { sockets } = createRttHarness(['session-a', 'session-b']);
-    const group = createGroupSnapshot('room-1', ['session-a', 'session-b']);
-    groupStateSnapshotsRepository.setGroupStateSnapshot(group);
-
-    await dispatchRtt(
-      sockets.get('session-a')!,
-      'session-a',
-      {
-        sessionIdFrom: 'session-a',
-        sessionIdTo: 'session-b',
-        rttMs: 0,
-        createdAtEpochMs: 1,
-        version: 1,
-      },
-      group,
-    );
-
-    expect(latestRttById().read('session-a::session-b')).toBeUndefined();
-  });
-
-  it('rejects RTT measurements without a shared active group', async () => {
-    configureTestCacheRepositories();
-    const { sockets } = createRttHarness(['session-a', 'session-b', 'session-c']);
-    const group = createGroupSnapshot('room-1', ['session-a', 'session-c']);
-    groupStateSnapshotsRepository.setGroupStateSnapshot(group);
-
-    await dispatchRtt(
-      sockets.get('session-a')!,
-      'session-a',
-      {
-        sessionIdFrom: 'session-a',
-        sessionIdTo: 'session-b',
-        rttMs: 12,
-        createdAtEpochMs: 1,
-        version: 1,
-      },
-      group,
-    );
-
-    expect(latestRttById().read('session-a::session-b')).toBeUndefined();
-  });
-
-  it('rejects RTT measurements outside the reporting edge policy', async () => {
-    configureTestCacheRepositories();
-    const { sockets, topologyService } = createRttHarness(
-      ['session-a', 'session-b', 'session-c', 'session-d'],
-      {
-        rtcTopologyOptions: {
-          rttReportingDegreeLimit: 1,
-        },
-      },
-    );
-    const group = createGroupSnapshot('room-1', [
-      'session-a',
-      'session-b',
-      'session-c',
-      'session-d',
-    ]);
-    groupStateSnapshotsRepository.setGroupStateSnapshot(group);
-    topologyService.updateGroupTopology(group);
-
-    await dispatchRtt(
-      sockets.get('session-c')!,
-      'session-c',
-      {
-        sessionIdFrom: 'session-c',
-        sessionIdTo: 'session-d',
-        rttMs: 12,
-        createdAtEpochMs: 1,
-        version: 1,
-      },
-      group,
-    );
-
-    expect(latestRttById().read('session-c::session-d')).toBeUndefined();
-  });
-
-  it('ignores stale RTT measurements before Vivaldi or topology work', async () => {
-    configureTestCacheRepositories();
-    const { sockets, topologyService } = createRttHarness(['session-a', 'session-b'], {
-      rtcTopologyOptions: {
-        rttRebuildDebounceMs: 0,
-      },
-    });
-    const group = createGroupSnapshot('room-1', ['session-a', 'session-b']);
-    groupStateSnapshotsRepository.setGroupStateSnapshot(group);
-    const observeRtt = vi.spyOn(vivaldiService, 'observeRtt');
-    const queueRttTopologyUpdate = vi.spyOn(topologyService, 'queueRttTopologyUpdate');
-
-    await dispatchRtt(
-      sockets.get('session-a')!,
-      'session-a',
-      {
-        sessionIdFrom: 'session-a',
-        sessionIdTo: 'session-b',
-        rttMs: 12,
-        createdAtEpochMs: 2,
-        version: 2,
-      },
-      group,
-    );
-
-    observeRtt.mockClear();
-    queueRttTopologyUpdate.mockClear();
-
-    await dispatchRtt(
-      sockets.get('session-a')!,
-      'session-a',
-      {
-        sessionIdFrom: 'session-a',
-        sessionIdTo: 'session-b',
-        rttMs: 4,
-        createdAtEpochMs: 1,
-        version: 1,
-      },
-      group,
-    );
-
-    expect(latestRttById().read('session-a::session-b')?.rttMs).toBe(12);
-    expect(observeRtt).not.toHaveBeenCalled();
-    expect(queueRttTopologyUpdate).not.toHaveBeenCalled();
-  });
-
-  it('rejects RTT measurements that would exceed the reporting degree', async () => {
-    configureTestCacheRepositories();
-    const { sockets } = createRttHarness(['session-a', 'session-b', 'session-c'], {
-      rtcTopologyOptions: {
-        rttReportingDegreeLimit: 1,
-      },
-    });
-    const group = createGroupSnapshot('room-1', ['session-a', 'session-b', 'session-c']);
-    groupStateSnapshotsRepository.setGroupStateSnapshot(group);
-
-    await dispatchRtt(
-      sockets.get('session-a')!,
-      'session-a',
-      {
-        sessionIdFrom: 'session-a',
-        sessionIdTo: 'session-b',
-        rttMs: 12,
-        createdAtEpochMs: 1,
-        version: 1,
-      },
-      group,
-    );
-    await dispatchRtt(
-      sockets.get('session-a')!,
-      'session-a',
-      {
-        sessionIdFrom: 'session-a',
-        sessionIdTo: 'session-c',
-        rttMs: 13,
-        createdAtEpochMs: 2,
-        version: 2,
-      },
-      group,
-    );
-
-    expect(latestRttById().read('session-a::session-b')).toBeDefined();
-    expect(latestRttById().read('session-a::session-c')).toBeUndefined();
-  });
-
-  it('rejects over-degree RTT measurements across active groups', async () => {
-    configureTestCacheRepositories();
-    const { sockets } = createRttHarness(['session-a', 'session-b', 'session-c'], {
-      rtcTopologyOptions: {
-        rttReportingDegreeLimit: 1,
-      },
-    });
-    const groupOne = createGroupSnapshot('room-1', ['session-a', 'session-b']);
-    const groupTwo = createGroupSnapshot('room-2', ['session-a', 'session-c']);
-    groupStateSnapshotsRepository.setGroupStateSnapshot(groupOne);
-    groupStateSnapshotsRepository.setGroupStateSnapshot(groupTwo);
-
-    await dispatchRtt(
-      sockets.get('session-a')!,
-      'session-a',
-      {
-        sessionIdFrom: 'session-a',
-        sessionIdTo: 'session-b',
-        rttMs: 12,
-        createdAtEpochMs: 1,
-        version: 1,
-      },
-      groupOne,
-    );
-    await dispatchRtt(
-      sockets.get('session-a')!,
-      'session-a',
-      {
-        sessionIdFrom: 'session-a',
-        sessionIdTo: 'session-c',
-        rttMs: 13,
-        createdAtEpochMs: 2,
-        version: 2,
-      },
-      groupTwo,
-    );
-
-    expect(latestRttById().read('session-a::session-b')).toBeDefined();
-    expect(latestRttById().read('session-a::session-c')).toBeUndefined();
-  });
-
-  it('does not schedule process-local topology work after RTT messages', async () => {
-    vi.useFakeTimers();
-    vi.setSystemTime(1_000);
-
-    try {
-      configureTestCacheRepositories();
-
-      const server = new JsonWebSocketServer();
-      const sockets = createSockets([
-        'session-a',
-        'session-b',
-        'session-c',
-        'session-d',
-        'session-e',
-      ]);
-
-      for (const [sessionId, socket] of sockets) {
-        server.addConnection(new ConnectionContext(sessionId, socket as never));
-      }
-
-      const service = new WsQueueBoxServerService(
-        new InMemoryQueueBox(new Map()),
-        new InMemoryQueueBox(new Map()),
-        server,
-        'server-1',
-      );
-      initRallarSystemWsTopics(service, {
-        rtcTopologyOptions: {
-          rttRebuildDebounceMs: 100,
-        },
-      });
-
-      const group = createGroupSnapshot('room-1', [...sockets.keys()]);
-      clientStateSnapshotsRepository.setClientStateSnapshots(
-        [...sockets.keys()].map(createClientSnapshot),
-      );
-      const senderSocket = sockets.get('session-a')!;
-
-      await senderSocket.dispatchMessage(
-        newALBroadcastMessage(
-          'session-a',
-          newALEventRoute(AppTopics.groupStateSnapshot, group.group.groupId, 'group-snapshot-1'),
-          'room',
-          AppTopics.groupStateSnapshot,
-          group,
-          {
-            groupRef: group.group,
-          },
-        ),
-      );
-
-      for (const socket of sockets.values()) {
-        socket.sent.length = 0;
-      }
-
-      const fullSnapshotScan = vi.spyOn(groupStateSnapshotsRepository, 'getAllGroupStateSnapshots');
-
-      for (const rtt of createCentralRttMeasurements([...sockets.keys()], 'session-a')) {
-        await senderSocket.dispatchMessage(
-          newALBroadcastMessage(
-            'session-a',
-            newALEventRoute(AppTopics.rtt, group.group.groupId, `rtt-${rtt.version}`),
-            'room',
-            AppTopics.rtt,
-            rtt,
-            {
-              groupRef: group.group,
-            },
-          ),
-        );
-      }
-
-      expect(fullSnapshotScan).not.toHaveBeenCalled();
-      fullSnapshotScan.mockRestore();
-      expect(countSentTopologyMessages(sockets)).toBe(0);
-
-      await vi.advanceTimersByTimeAsync(99);
-      expect(countSentTopologyMessages(sockets)).toBe(0);
-
-      await vi.advanceTimersByTimeAsync(1);
-      expect(countSentTopologyMessages(sockets)).toBe(0);
-    } finally {
-      vi.useRealTimers();
-    }
-  });
-
   it('does not create topology work from an inbound group snapshot when app outbox owns topology', async () => {
     configureTestCacheRepositories();
 
@@ -639,136 +299,6 @@ describe('Rallar system websocket topics RTC topology', () => {
     expect(countSentTopologyMessages(createSocketsFrom([senderSocket, peerSocket]))).toBe(0);
   });
 
-  it('routes RTT-triggered topology recomputes through app outbox ownership', async () => {
-    vi.useFakeTimers();
-    vi.setSystemTime(1_000);
-
-    try {
-      configureTestCacheRepositories();
-
-      const server = new JsonWebSocketServer();
-      const sockets = createSockets([
-        'session-a',
-        'session-b',
-        'session-c',
-        'session-d',
-        'session-e',
-      ]);
-
-      for (const [sessionId, socket] of sockets) {
-        server.addConnection(new ConnectionContext(sessionId, socket as never));
-      }
-
-      const appOutboxQueue = new InMemoryQueueBox(new Map());
-      const runtimeRepository = new FakeRuntimeStateRepository();
-      const outboxQueueReader = new OutboxQueueReader(appOutboxQueue);
-      const wake = vi.fn();
-      const service = new WsQueueBoxServerService(
-        new InMemoryQueueBox(new Map()),
-        new InMemoryQueueBox(new Map()),
-        server,
-        'server-1',
-      );
-      const group = createGroupSnapshot('room-1', [...sockets.keys()]);
-      const findGroupSnapshotByRef = vi.fn(() => Promise.resolve(group));
-      initRallarSystemWsTopics(service, {
-        rtcTopologyOptions: {
-          rttRebuildDebounceMs: 100,
-        },
-        rtcTopologyManagement: createTopologyManagement(undefined, findGroupSnapshotByRef),
-        rtcTopologyAppOutbox: {
-          outboxQueueReader,
-          ...createTopologyExecutionDependencies(runtimeRepository),
-          wake,
-          findGroupSnapshotByRef,
-        },
-      });
-
-      clientStateSnapshotsRepository.setClientStateSnapshots(
-        [...sockets.keys()].map(createClientSnapshot),
-      );
-      const senderSocket = sockets.get('session-a')!;
-
-      await senderSocket.dispatchMessage(
-        newALBroadcastMessage(
-          'session-a',
-          newALEventRoute(AppTopics.groupStateSnapshot, group.group.groupId, 'group-snapshot-1'),
-          'room',
-          AppTopics.groupStateSnapshot,
-          group,
-          {
-            groupRef: group.group,
-          },
-        ),
-      );
-      expect(await appOutboxQueue.getAllKeys()).toEqual([]);
-      groupStateSnapshotsRepository.removeGroupStateSnapshotByRef(group.group);
-
-      const fullSnapshotScan = vi.spyOn(groupStateSnapshotsRepository, 'getAllGroupStateSnapshots');
-
-      const rtt = {
-        sessionIdFrom: 'session-a',
-        sessionIdTo: 'session-b',
-        rttMs: 12,
-        createdAtEpochMs: 1_000,
-        version: 1,
-      };
-      await senderSocket.dispatchMessage(
-        newALBroadcastMessage(
-          'session-a',
-          newALEventRoute(AppTopics.rtt, group.group.groupId, `rtt-${rtt.version}`),
-          'room',
-          AppTopics.rtt,
-          rtt,
-          { groupRef: group.group },
-        ),
-      );
-      groupStateSnapshotsRepository.setGroupStateSnapshot(group);
-
-      expect(fullSnapshotScan).not.toHaveBeenCalled();
-      fullSnapshotScan.mockRestore();
-      expect(wake).toHaveBeenCalled();
-      expect(findGroupSnapshotByRef).toHaveBeenCalledWith({
-        applicationId: group.group.applicationId,
-        workspaceId: group.group.workspaceId,
-        groupId: group.group.groupId,
-      });
-      expect(countSentTopologyMessages(sockets)).toBe(0);
-      const [key] = await appOutboxQueue.getAllKeys();
-      expect(key).toMatchObject({
-        topicId: 'app-outbox.rtc-topology',
-        resourceId: expect.stringContaining('room-1'),
-      });
-      const entry = await appOutboxQueue.getItem(key!);
-      const message = JSON.parse(entry!.resource) as ALMessage;
-      const envelope = JSON.parse(message.payload.resource) as {
-        resourceId: string;
-        contextId: string;
-        data: {
-          kind: string;
-          groupSnapshot: GroupSnapshot;
-          requestedRttVersion: number;
-          requestOptions: object;
-          publish: boolean;
-        };
-      };
-      expect(message.route).toEqual(key);
-      expect(envelope).toMatchObject({
-        resourceId: expect.any(String),
-        contextId: expect.stringContaining('group=room-1'),
-        data: {
-          kind: 'rtt-refresh',
-          groupSnapshot: group,
-          requestedRttVersion: rtt.version,
-          requestOptions: {},
-          publish: true,
-        },
-      });
-    } finally {
-      vi.useRealTimers();
-    }
-  });
-
   it('converges multiple app-outbox publishers on one immutable work identity', async () => {
     configureTestCacheRepositories();
     const appOutboxQueue = new InMemoryQueueBox(new Map());
@@ -849,72 +379,12 @@ class FakeSocket {
     }
   }
 }
-
 function createSockets(sessionIds: readonly string[]): Map<string, FakeSocket> {
   return new Map(sessionIds.map((sessionId) => [sessionId, new FakeSocket()]));
 }
 
 function createSocketsFrom(sockets: readonly FakeSocket[]): Map<string, FakeSocket> {
   return new Map(sockets.map((socket, index) => [`socket-${index}`, socket]));
-}
-
-function createRttHarness(
-  sessionIds: readonly string[],
-  options: Readonly<{
-    rtcTopologyOptions?: ConstructorParameters<typeof RallarRtcTopologyService>[0];
-    runtimeRepository?: FakeRuntimeStateRepository;
-    enqueueRtcRttMutation?: InitRallarSystemWsTopicsOptions['enqueueRtcRttMutation'];
-  }> = {},
-): {
-  readonly sockets: Map<string, FakeSocket>;
-  readonly topologyService: RallarRtcTopologyService;
-} {
-  const server = new JsonWebSocketServer();
-  const sockets = createSockets(sessionIds);
-  for (const [sessionId, socket] of sockets) {
-    server.addConnection(new ConnectionContext(sessionId, socket as never));
-  }
-
-  const service = new WsQueueBoxServerService(
-    new InMemoryQueueBox(new Map()),
-    new InMemoryQueueBox(new Map()),
-    server,
-    'server-1',
-  );
-  const topologyService = new RallarRtcTopologyService(options.rtcTopologyOptions);
-  initRallarSystemWsTopics(service, {
-    rtcTopologyService: topologyService,
-    ...(options.runtimeRepository
-      ? { rtcTopologyRuntimeState: { repository: options.runtimeRepository } }
-      : {}),
-    enqueueRtcRttMutation: options.enqueueRtcRttMutation,
-  });
-
-  return { sockets, topologyService };
-}
-
-async function dispatchRtt(
-  socket: FakeSocket,
-  senderId: string,
-  rtt: Readonly<{
-    sessionIdFrom: string;
-    sessionIdTo: string;
-    rttMs: number;
-    createdAtEpochMs: number;
-    version: number;
-  }>,
-  group?: GroupSnapshot,
-): Promise<void> {
-  await socket.dispatchMessage(
-    newALBroadcastMessage(
-      senderId,
-      newALEventRoute(AppTopics.rtt, group?.group.groupId ?? 'room-1', `rtt-${rtt.version}`),
-      'room',
-      AppTopics.rtt,
-      rtt,
-      group ? { groupRef: group.group } : undefined,
-    ),
-  );
 }
 
 function countSentTopologyMessages(sockets: ReadonlyMap<string, FakeSocket>): number {
@@ -948,36 +418,6 @@ function createUnusedDatabase(): PSqlSql {
   database.begin = () =>
     Promise.reject(new Error('Unexpected transaction in WS routing unit test'));
   return database;
-}
-
-function createCentralRttMeasurements(
-  sessionIds: readonly string[],
-  centralSessionId: string,
-): readonly {
-  readonly sessionIdFrom: string;
-  readonly sessionIdTo: string;
-  readonly rttMs: number;
-  readonly createdAtEpochMs: number;
-  readonly version: number;
-}[] {
-  const measurements = [];
-  let version = 1;
-
-  for (let i = 0; i < sessionIds.length; i++) {
-    for (let j = i + 1; j < sessionIds.length; j++) {
-      const from = sessionIds[i];
-      const to = sessionIds[j];
-      measurements.push({
-        sessionIdFrom: from,
-        sessionIdTo: to,
-        rttMs: from === centralSessionId || to === centralSessionId ? 1 : 100,
-        createdAtEpochMs: version,
-        version: version++,
-      });
-    }
-  }
-
-  return measurements;
 }
 
 function createResilience(): ResilienceDto {
